+++ CHANGELOG.md
@@ -5,6 +5,7 @@ XXXX-XX-XX, v1.0.1
 ------------------
 
 - fixed [#336](https://github.com/n1k0/casperjs/issues/336) - Test result duration may have an exotic value
+- Added `casper.mouse.doubleclick()`
 
 2012-12-24, v1.0.0
 ------------------